Fix rec uploads for alma-9/perl-5.32
On installing this to the new alma-9 server from the old centos-7
server, the $req->upload function was throwing an 'internal apreq error'
if there was no upload
(https://www.inter-corporate.com/kb/internal-apreq-error.pl).

But the Apache2::Request object was also refusing to process any of the
form elements after the empty upload.

Fix is to remove the "file" input element if they're not uploading a new
recording.